change LMR reduction with higher precision

+++ src/search.cpp
@@ -1147,7 +1147,7 @@ moves_loop: // When in check, search starts here
               || (cutNode && (ss-1)->moveCount > 1)))
       {
           Depth r = reduction(improving, depth, moveCount, delta, thisThread->rootDelta);
-          double deltaR = 0;
+          int deltaR = 0;
           // Decrease reduction at some PvNodes (~2 Elo)
           if (   PvNode
               && bestMoveCount <= 3)
